backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
python-dotenv==1.0.1
google-genai==1.14.0
pandas==2.2.2
Pillow==10.4.0
pdf2image==1.17.0
